requirements.txt
# Core
python-dotenv==1.0.1

# Synthetic data generation (Sprint 0–1)
faker==24.9.0
numpy==1.26.4
pandas==2.2.2

# Database
# SQLite is built into Python — no extra package needed

# Web UI (Sprint 3+)
streamlit==1.33.0
plotly==5.22.0

# LLM integration (Sprint 3+)
openai==1.30.1
anthropic==0.28.0

# HTTP mocking for demo data sources (Sprint 2)
responses==0.25.3
httpx==0.27.0

# ML (Sprint 5)
scikit-learn==1.4.2
shap==0.45.1
joblib==1.4.2

# NLP / sentiment (Sprint 6)
transformers==4.41.1
torch==2.3.0

# Email (Sprint 4 — demo mode uses logs only)
sendgrid==6.11.0

# PDF / contracts (Sprint 7)
reportlab==4.2.0

# Testing
pytest==8.2.0
pytest-cov==5.0.0

# Utilities
python-dateutil==2.9.0
rich==13.7.1
